res removed uses of strncpy(). (strncpy() does more than you expect   [it fills the destination buffer up to then end with 0] and less   than you want [it doesn't guarantee null termination].) res fixed some signed vs unsigned warnings.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@21046 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/cstool/framedataholder.h
@@ -59,7 +59,7 @@ public:
    */
   T& GetUnusedData (bool& created, uint frameNumber)
   {
-    if ((clearReq != ~0) && (clearReq != frameNumber))
+    if ((clearReq != (uint)~0) && (clearReq != frameNumber))
       data.DeleteAll();
     lastFrame = frameNumber;
     created = false;
